Fix MPS inference: float16 resize crash + alpha video detection
autocast float16 outputs crash cv2.resize(INTER_LANCZOS4) — cast to
float32 before numpy. AlphaHint dir with video files (no images) was
silently producing 0-frame runs — add video fallback + frame count
logging/skip guard. Tested on M3 Max.

+++ CorridorKeyModule/inference_engine.py
@@ -168,8 +168,8 @@ class CorridorKeyEngine:
 
         # 6. Post-Process (Resize Back to Original Resolution)
         # We use Lanczos4 for high-quality resampling to minimize blur when going back to 4K/Original.
-        res_alpha = pred_alpha[0].permute(1, 2, 0).cpu().numpy()
-        res_fg = pred_fg[0].permute(1, 2, 0).cpu().numpy()
+        res_alpha = pred_alpha[0].permute(1, 2, 0).float().cpu().numpy()
+        res_fg = pred_fg[0].permute(1, 2, 0).float().cpu().numpy()
         res_alpha = cv2.resize(res_alpha, (w, h), interpolation=cv2.INTER_LANCZOS4)
         res_fg = cv2.resize(res_fg, (w, h), interpolation=cv2.INTER_LANCZOS4)
